refactor(src): improve time parsing and refactor code style in adapters
- Refactored WebSocket message parsing in DHLM adapter to simplify control flow
- Updated file path handling in `load_dump_file` to use `Option<&PathBuf>` instead of `&Option<PathBuf>`
- Changed NLS adapter to compute lap centiseconds with rounding via third fractional digit
- Replaced `const fn` date conversions in Insights with `fn`, improving safety with `i32::try_from`
- Normalized driver name token logic to lowercase only after first alphabetic character
- Reformatted imports across multiple modules for consistent style and readability

+++ src/web/server.rs
@@ -175,7 +175,7 @@ pub async fn run(mode: RunMode) -> Result<(), Box<dyn std::error::Error>> {
 
    if mode == RunMode::DaemonChild {
       write_runtime_info(&runtime_info)?;
-      write_pid(runtime_info.pid as i32)?;
+      write_pid(runtime_info.pid)?;
    }
 
    let shutdown = async move {
